backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
chromadb
pymupdf
groq
python-dotenv
openai
pydantic-settings
pydantic
python-multipart
httpx
jinja2
func_timeout
pandas
openpyxl
python-docx
reportlab
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
email-validator
google-auth
